django-pjt/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==4.2.13
django-allauth==0.61.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-seed==0.3.1
djangorestframework==3.15.1
exceptiongroup==1.2.1
executing==2.0.1
Faker==25.2.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httplib2==0.22.0
idna==3.7
ipython==8.18.1
jedi==0.19.1
lxml==5.2.2
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.0
parso==0.8.4
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
stack-data==0.6.3
toposort==1.10
tqdm==4.66.4
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
webdriver-manager==4.0.1
wsproto==1.2.0
